+++ .dev.vars
@@ -1,0 +1,1 @@
+NEXTJS_ENV=development
