+++ .env
@@ -52,3 +52,8 @@ GC_BUCKET_NAME=openbeta-staging
 # A comma-separated-list of test area IDs
 # to exclude from the feed
 NEXT_PUBLIC_TEST_AREA_IDS=18c5dd5c-8186-50b6-8a60-ae2948c548d1
+
+
+OPENNEXT_PLATFORM="cloudflare"
+NEXT_PUBLIC_CLOUDFLARE=true
+NEXT_PUBLIC_AWS=false
