+++ open-next.config.ts
@@ -1,0 +1,26 @@
+// default open-next.config.ts file created by @opennextjs/cloudflare
+
+import cache from "@opennextjs/cloudflare/kv-cache";
+
+const config = {
+  default: {
+    override: {
+      wrapper: "cloudflare-node",
+      converter: "edge",
+      incrementalCache: async () => cache,
+      tagCache: "dummy",
+      queue: "dummy",
+    },
+  },
+
+  middleware: {
+    external: true,
+    override: {
+      wrapper: "cloudflare-edge",
+      converter: "edge",
+      proxyExternalRequest: "fetch",
+    },
+  },
+};
+
+export default config;
