+++ src/app/(default)/components/Topo/Topo.tsx
@@ -1,3 +1,4 @@
+'use client'
 import { useCallback, useEffect } from 'react'
 import { RouteInfo, clearProject, drawTopo, initPaper, zoomViaWheel } from './paperFunctions'
