+++ src/app/api/mobile/login/route.ts
@@ -2,6 +2,7 @@ import { NextRequest, NextResponse } from 'next/server'
 import * as Auth0 from 'auth0'
 import { auth0Client, isNullOrEmpty } from '@/js/auth/mobile'
 import { withMobileAuth } from '@/js/auth/withMobileAuth'
+import { errorHandler } from './errorHandler'
 
 /**
  * Mobile login handler
@@ -37,14 +38,3 @@ async function postHandler (request: NextRequest): Promise<NextResponse> {
 }
 
 export const POST = withMobileAuth(postHandler)
-
-/**
- * Handle Auth0 errors
- */
-export const errorHandler = (error: any): NextResponse => {
-  console.error('#### Auth0 error ####', error)
-  if (error instanceof Auth0.AuthApiError) {
-    return NextResponse.json({ error: error?.error_description ?? '' }, { status: error?.statusCode ?? 401 })
-  }
-  return NextResponse.json({ error: 'Unexpected auth error' }, { status: 401 })
-}
