+++ src/pages/api/withAuth.ts
@@ -4,7 +4,7 @@
 */
 import { NextApiHandler } from 'next'
 import { getServerSession } from 'next-auth'
-import { authOptions } from '@/app/api/auth/[...nextauth]/route'
+import { authOptions } from '@/app/api/auth/[...nextauth]/authOptions'
 
 const withAuth = (handler: NextApiHandler): NextApiHandler => {
   return async (req, res) => {
