+++ wrangler.jsonc
@@ -1,0 +1,14 @@
+{
+    "name": "open-tacos",
+    "compatibility_date": "2025-02-16",
+    "compatibility_flags": [
+      "nodejs_compat"
+    ],
+    "pages_build_output_dir": ".vercel/output/static",
+    "$schema": "node_modules/wrangler/config-schema.json",
+    "main": ".open-next/worker.js"
+    // "assets": {
+    //     "directory": ".open-next/assets",
+    //     "binding": "ASSETS"
+    // }
+  }
